+++ main.py
@@ -1,6 +1,8 @@
 import asyncio
 import os
 from dotenv import load_dotenv
+
+from utils import convert_to_json_data
 from video_analyser import analyse_video
 from spider import download_video
 
@@ -8,8 +10,8 @@ load_dotenv()
 API_KEY = os.getenv("KEY")
 
 
-def download_and_analyse_video(url, csv_path, transcript_path, api_key):
-    video_path = download_video(url)
+def download_and_analyse_video(url, csv_path, transcript_path, api_key, delete_temp=True):
+    video_path, video_id = download_video(url)
     asyncio.run(
         analyse_video(
             video_path,
@@ -18,16 +20,23 @@ def download_and_analyse_video(url, csv_path, transcript_path, api_key):
             api_key=api_key,
         )
     )
-    os.remove(video_path)
-    return csv_path, transcript_path
+
+    json_result = convert_to_json_data(csv_path, transcript_path, video_id)
+    if delete_temp:
+        os.remove(video_path)
+        os.remove(csv_path)
+        os.remove(transcript_path)
+
+    return json_result
 
 
 if __name__ == "__main__":
-    asyncio.run(
-        analyse_video(
-            video_path="test/test.mp4",
-            csv_path="results/test.csv",
-            transcript_path="results/test.txt",
-            api_key=API_KEY,
-        )
-    )
+    csv_path, transcript_path = asyncio.run(analyse_video(
+        video_path="test/test.mp4",
+        csv_path="results/test.csv",
+        transcript_path="results/test.txt",
+        api_key=API_KEY,
+        debug=True,
+    ))
+    json_result = convert_to_json_data(csv_path, transcript_path)
+    print(json_result)
